(PC-10707)[API] refactor: accessibility mixin for Offer and Venue
Offer already has an accessibility mixin: this commit moves it to a more
generic module in order to be used by Venue.
