+++ api/src/pcapi/models/accessibility_mixin.py
@@ -1,0 +1,11 @@
+import sqlalchemy as sa
+
+
+class AccessibilityMixin:
+    audioDisabilityCompliant = sa.Column(sa.Boolean, nullable=True)
+
+    mentalDisabilityCompliant = sa.Column(sa.Boolean, nullable=True)
+
+    motorDisabilityCompliant = sa.Column(sa.Boolean, nullable=True)
+
+    visualDisabilityCompliant = sa.Column(sa.Boolean, nullable=True)
